rkaiq: Release v4.8.7
Updated CLs:
904c5c3d IQ: close af as default
d97bfd5b AF: some small modification
13a04de3 CAC: Support IQ Tool
2e809296 awb: fix some bugs
ff20dc4e Kernel headers: Sync with driver
cb128838 aiq: gic: update for isp32
6bdc9184 blc:fix some bug
e29797c4 ynrv22: fix ui bug
c3b009b3 M4: add new macro for tuning TOOL
2a071487 algo: ldchV21: add uapi & sample support
fd9f09ea Amerge: update log setting
85546e4d AF: change M4 description
faf69f96 Adehaze: Fix bringup bug
3f54bdcb AE: fix ae stats bug
23a69ede Fix blc params error in FakeCamera0.json
ec62d533 Adrc: Fix some bugs
53286353 nr supprot for tuning tool
81f91d9e ABLCV32 & SharpV33: fix the bug
87d5379d CCM: fix some bug
be6b8f2b AE & Debayer:modify for RKTool
f3f60255 AE: ae stats adapt for ob_offset and awb1_gain
c70b9c40 changes for RK Tool   1.update adrc head   2.update merge
params in json
5c362d71 Adrc: fix gas_l0 bug caused by
dfd11a3dfba10cfdb9fe08b659ca742e5a14266a
a1ed7dd6 Adehaze: support ISO as ctrl data
6d3bdc59 CCM: simplify code
48081368 Adehaze: support local gain setting
619e7e32 algo: ldchV21: support ldch on the rv1106
72487181 AE: modify exp_i2c_params position
bf39e869 AE:simplify codes and procRes
e535a98a AF: reduce memory
74db2726 AF: change sobel_sel value
0753e371 rv1106 iq: FakeCamere0.json sync with html os04a10 params
9ec22bbd aiq: Fix first frame sof wrong in offline mode
0ef36df1 lsc: fix lsc result convert.
66b1783d Adrc: update log setting
0b757446 json: comfirm four modifications below
817fd3bb AE: modify params in sc500ai/sc4336 json
a72a0241 Revert "AE: modify delay scheme"
b7f3c48c RkAiqxxxHandle: fix wrong 3a result caused by no 3a stats
dfd11a3d Adrc: add clip function for gas_lx
5ed99006 Debayer:modify group mode
f3b4f6ca Fix update drc params in isp32 json
4c18c98d json: add sc4336 json for 1106
20ae3abb json:sc500ai 3588 json which need to be changed for 1106
98dd829c Release v4.0x8.6 for rv1106 bringup
782a683a blcV32: modify blc1*predgain case


Former-commit-id: 0226223b535274aff7ab2ac836ae15685d85b0ff

+++ rkaiq/aiq_core/RkAiqCore.h
@@ -425,7 +425,7 @@ public:
 
     void post_message (SmartPtr<XCamMessage> &msg);
     int32_t getGroupId(RkAiqAlgoType_t type);
-    XCamReturn getGroupSharedParams(int32_t groupId, RkAiqAlgosGroupShared_t* &shared);
+    XCamReturn getGroupSharedParams(uint64_t groupId, RkAiqAlgosGroupShared_t* &shared);
     uint64_t getCustomEnAlgosMask() {
         return mCustomEnAlgosMask;
     }
